See description
- Autoloading from PSR-0 to PSR-4
- PHP CS fixer used to be compliant with PSR-1 and PSR-2
- HHVM tests added
- HHVM Badge added
- Versions and download badges added
- Fixed tests failing
- Scrutinizer configuration file added (to replace PHPMD and add new
quality code analyzers).
- Bugs fixed
