+++ .gitignore
@@ -1,5 +1,2 @@
-
-.svn
-/.buildpath
-/.project
-/.settings
+vendor/
+build/
