+++ .scrutinizer.yml
@@ -1,0 +1,14 @@
+before_commands:
+    - "composer install --prefer-dist"
+
+tools:
+    php_mess_detector: true
+    php_code_sniffer: true
+    php_analyzer: true
+    sensiolabs_security_checker: true
+    php_code_coverage: true
+    php_cpd: true
+    php_pdepend:
+        excluded_dirs: [vendor/, tests/*]
+filter:
+    excluded_paths: [vendor/]
