+++ .travis.yml
@@ -5,8 +5,15 @@ php:
   - 5.4
   - 5.5
   - 5.6
+  - hhvm
 
 before_script:
   - composer install
 
-script: phpunit --coverage-text
+before_script:
+    - curl -s http://getcomposer.org/installer | php
+    - php composer.phar update --dev --no-interaction
+    - mkdir -p build/logs
+
+script:
+    - php vendor/bin/phpunit --coverage-text
