+++ README.md
@@ -4,29 +4,29 @@ base32
 Base32 Encoder/Decoder for PHP according to RFC 4648
 
 [![Build Status](https://secure.travis-ci.org/ChristianRiesen/base32.png)](http://travis-ci.org/ChristianRiesen/base32)
+[![HHVM Status](http://hhvm.h4cc.de/badge/christian-riesen/base32.png)](http://hhvm.h4cc.de/package/christian-riesen/base32)
 
-Did you like this? Flattr it:
+[![Latest Stable Version](https://poser.pugx.org/christian-riesen/base32/v/stable.png)](https://packagist.org/packages/christian-riesen/base32) [![Total Downloads](https://poser.pugx.org/christian-riesen/base32/downloads.png)](https://packagist.org/packages/christian-riesen/base32) [![Latest Unstable Version](https://poser.pugx.org/christian-riesen/base32/v/unstable.png)](https://packagist.org/packages/christian-riesen/base32) [![License](https://poser.pugx.org/christian-riesen/base32/license.png)](https://packagist.org/packages/christian-riesen/base32)
+
+Do you like this? Flattr it:
 
 [![Flattr base32](http://api.flattr.com/button/flattr-badge-large.png)](http://flattr.com/thing/720563/ChristianRiesenbase32-on-GitHub)
 
 Usage
 -----
 
-```php
-<?php
+    <?php
 
-// Include class or user autoloader
-use Base32\Base32;
+    // Include class or user autoloader
+    use Base32\Base32;
 
-$string = 'fooba';
+    $string = 'fooba';
 
-$encoded = Base32::encode($string);
-// $encoded contains now 'MZXW6YTB'
+    $encoded = Base32::encode($string);
+    // $encoded contains now 'MZXW6YTB'
 
-$decoded = Base32::decode($encoded);
-// $decoded is again 'fooba'
-
-```
+    $decoded = Base32::decode($encoded);
+    // $decoded is again 'fooba'
 
 
 About
