+++ composer.json
@@ -1,7 +1,7 @@
 {
     "name": "christian-riesen/base32",
     "type": "library",
-    "description": "Base32 encoder/decoder, unit tested, according to RFC 4648",
+    "description": "Base32 encoder/decoder according to RFC 4648",
     "keywords": ["base32","encode","decode","rfc4648"],
     "homepage": "https://github.com/ChristianRiesen/base32",
     "license": "MIT",
@@ -16,9 +16,18 @@
     "require": {
         "php": ">=5.3.0"
     },
+    "require-dev": {
+        "phpunit/phpunit": "4.*",
+        "satooshi/php-coveralls": "0.*"
+    },
 	"autoload": {
-        "psr-0": {
-            "Base32": "src"
+        "psr-4": {
+            "Base32\\": "src/"
+        }
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "1.1.x-dev"
         }
     }
 }
