+++ phpunit.xml.dist
@@ -1,15 +1,24 @@
 <?xml version="1.0" encoding="UTF-8"?>
-
-<phpunit colors="true">
+<phpunit
+    backupGlobals="false"
+    backupStaticAttributes="false"
+    convertErrorsToExceptions="true"
+    convertNoticesToExceptions="true"
+    convertWarningsToExceptions="true"
+    processIsolation="false"
+    stopOnFailure="false"
+    syntaxCheck="false"
+    bootstrap="tests/bootstrap.php"
+    colors="true">
     <testsuites>
         <testsuite name="Base32 Test Suite">
-            <directory>tests/Base32/</directory>
+            <directory suffix="Test.php">tests/</directory>
         </testsuite>
     </testsuites>
 
     <filter>
         <whitelist>
-            <directory suffix=".php">src/Base32/</directory>
+            <directory suffix=".php">src/</directory>
         </whitelist>
     </filter>
