+++ tests/bootstrap.php
@@ -1,0 +1,5 @@
+<?php
+
+$loader = require __DIR__ . '/../vendor/autoload.php';
+$loader->add("Base32", __DIR__);
+$loader->register();
